3번까지 통과

+++ src/main/java/com/sparta/mydelivery/dto/RestaurantsDto.java
@@ -9,6 +9,6 @@ import lombok.Setter;
 @NoArgsConstructor
 public class RestaurantsDto {
     private String name;
-    private Long minOrderPrice;
-    private Long deliveryFee;
+    private int minOrderPrice;
+    private int deliveryFee;
 }
